Refactor code structure
Refactor the existing codebase for improved organization and maintainability.
[skip gpt_engineer]
