+++ src/utils/fetchWithTimeout.ts
@@ -1,0 +1,33 @@
+export const FETCH_TIMEOUT = 600000; // 10 minutes timeout
+
+export const fetchWithTimeout = async (
+  url: string, 
+  options: RequestInit,
+  timeout: number
+) => {
+  const controller = new AbortController();
+  const id = setTimeout(() => controller.abort(), timeout);
+
+  try {
+    const response = await fetch(url, {
+      ...options,
+      signal: controller.signal
+    });
+    clearTimeout(id);
+    
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+    
+    return response;
+  } catch (error) {
+    clearTimeout(id);
+    if (error instanceof Error) {
+      if (error.name === 'AbortError') {
+        throw new Error('Request timed out');
+      }
+      throw error;
+    }
+    throw new Error('Network error occurred');
+  }
+};
