+++ src/utils/responseHandler.ts
@@ -1,0 +1,54 @@
+export const extractResponseContent = (data: any): string => {
+  if (typeof data === 'string') return data;
+  
+  if (Array.isArray(data)) {
+    const firstItem = data[0];
+    if (!firstItem) return "No response content available";
+    
+    if (typeof firstItem === 'string') return firstItem;
+    
+    if (typeof firstItem === 'object') {
+      if (firstItem.message?.content && firstItem.message?.role === 'assistant') {
+        return firstItem.message.content;
+      }
+      
+      const possibleContent = firstItem.message?.content || 
+                            firstItem.content ||
+                            firstItem.output ||
+                            firstItem.text ||
+                            firstItem.response;
+                            
+      if (possibleContent) return possibleContent;
+      
+      try {
+        return JSON.stringify(firstItem);
+      } catch (e) {
+        console.warn('Failed to stringify response:', e);
+        return "Unable to process response format";
+      }
+    }
+  }
+  
+  if (typeof data === 'object') {
+    if (data.message?.content && data.message?.role === 'assistant') {
+      return data.message.content;
+    }
+    
+    const possibleContent = data.message?.content || 
+                          data.content ||
+                          data.output ||
+                          data.text ||
+                          data.response;
+                          
+    if (possibleContent) return possibleContent;
+    
+    try {
+      return JSON.stringify(data);
+    } catch (e) {
+      console.warn('Failed to stringify response:', e);
+      return "Unable to process response format";
+    }
+  }
+  
+  return "Unexpected response format";
+};
